graph_services/requirements.txt
pytest>=6.2.0
Flask[async]>=2.3.2
Flask-SocketIO>=5.0.1
Flask-Cors>=3.0.10
Flask-RESTful>=0.3.8
networkx>=3.1
numpy>=1.21.2
pandas>=1.4.2
python-dotenv>=0.19.0
pylint>=2.16.1
black>=23.3.0
isort>=5.12.0
redis[hiredis]>=5.0.0b3
gevent>=21.1.2
gevent-websocket>=0.10.1
eventlet>=0.33.3